feat: Add support for how-to auto shown feature
This commit adds support for the how-to auto shown feature in the settings. It includes changes to the `ISettings` interface, `SettingsProvider`, `GameScreen`, and `HowToModal` widget. The `ISettings` interface now includes methods to get and save the how-to auto shown value. The `SettingsProvider` class has been updated with a new method to handle the how-to auto shown logic. In the `GameScreen`, a check has been added to show the how-to modal if it hasn't been auto shown yet. Finally, the `HowToModal` widget now includes an optional callback function that is called when the close button is pressed.

Closes #123