feat: Add support for displaying leaderboards on iOS
This commit adds the necessary code to display leaderboards on iOS devices. It imports the `dart:io` library and includes a conditional statement to check if the platform is iOS before showing the "Leaderboards" button. If the user is signed in, tapping the button will show the leaderboards using the specified leaderboard ID. If not signed in, a snackbar message prompts the user to sign in.