Add horizontal swipe functionality to BeltOfDestiny game
This commit adds the ability to perform horizontal swipes in the BeltOfDestiny game. The `BeltOfDestiny` class now implements the `HorizontalDragDetector` interface and includes methods for handling horizontal drag events. Swiping right toggles the direction of the control arm if it is not already straight down, while swiping left toggles the direction if it is straight down. The `_isSwiping` flag is used to track whether a swipe is currently in progress.